category synonym admin done

+++ src/main/java/com/tmoncorp/admin/controller/CategoryController.java
@@ -10,10 +10,7 @@ import com.tmoncorp.admin.service.UploadFileDecodeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 
 import java.text.ParseException;
@@ -28,6 +25,14 @@ import java.util.stream.Collectors;
 @RestController
 @RequestMapping(value = "/category")
 public class CategoryController {
+
+    private enum SynchronizeState {
+        CATEGORY_RENEWAL(0), UPLOAD(1), UPDATE(2);
+        private final int code;
+        SynchronizeState(int code) {
+            this.code = code;
+        }
+    }
     private SynonymCategoryRepository synonymCategoryRepository;
     private OriginalCategoryRepository originalCategoryRepository;
 
@@ -36,7 +41,6 @@ public class CategoryController {
         this.synonymCategoryRepository = synonymCategoryRepository;
         this.originalCategoryRepository = originalCategoryRepository;
     }
-
 
     @RequestMapping(value = "/", method = RequestMethod.GET)
     public ResponseEntity<Map> dataLoadFromDB() {
@@ -52,12 +56,15 @@ public class CategoryController {
         }
     }
 
+    // Renewal new category
     @RequestMapping(value = "/renewal", method = RequestMethod.GET)
     public ResponseEntity<Map> renewalCategory() {
         Map<String, Object> responseMap = new HashMap<>();
 
         try {
+            // Parse originalCategoryList as synonymCategoryList
             List<SynonymCategory> synonymCategoryList = new CategoryParseService(originalCategoryRepository.findAll()).getNewSynonymCategoryList();
+            // Update synonymCategoryList
             CategoryRenewalService categoryRenewalService = new CategoryRenewalService(synonymCategoryRepository.findAll(), synonymCategoryList);
 
             synonymCategoryList = categoryRenewalService.getRenewedSynonymCategoryList();
@@ -73,7 +80,7 @@ public class CategoryController {
         }
     }
 
-
+    // Upload .csv file to UI
     @RequestMapping(value = "/upload", method = RequestMethod.POST)
     public ResponseEntity<Map> uploadFile(MultipartHttpServletRequest request) {
         Map<String, Object> responseMap = new HashMap<>();
@@ -99,11 +106,13 @@ public class CategoryController {
         }
     }
 
+    // Download .csv file from html
     @RequestMapping(value = "/download", method = RequestMethod.GET)
     public ResponseEntity<Map> downloadAsCSV() {
         Map<String, Object> responseMap = new HashMap<>();
 
         try {
+            // using Stream, make csv string
             String csvFormatString = synonymCategoryRepository.findAll().stream()
                     .map(SynonymCategory::makeCsvString)
                     .collect(Collectors.joining(System.getProperty("line.separator")));
@@ -117,13 +126,30 @@ public class CategoryController {
         }
     }
 
+    // Synchronize UI & DB
     @RequestMapping(value = "/synchronize", method = RequestMethod.POST)
-    public ResponseEntity<Map> synchronizeWithDB(@RequestBody List<SynonymCategory> synonymCategoryList) {
+    public ResponseEntity<Map> synchronizeWithDB(@RequestBody List<SynonymCategory> synonymCategoryList, @RequestParam("renewal") int isRenewal,
+                                                 @RequestParam("changedCatNo") List<Integer> changedCatNo) {
         Map<String, Object> responseMap = new HashMap<>();
 
         try {
-            synonymCategoryRepository.deleteAll();
-            synonymCategoryRepository.save(synonymCategoryList);
+            // 카테고리 개편시 새로운 카테고리로 DB 업데이트
+            if (isRenewal == SynchronizeState.CATEGORY_RENEWAL.code) {
+                synonymCategoryRepository.deleteAll();
+                synonymCategoryRepository.save(synonymCategoryList);
+
+                // csv file upload
+            } else if (isRenewal == SynchronizeState.UPLOAD.code) {
+                synonymCategoryRepository.save(synonymCategoryList);
+
+                // 일반적인 DB 반영일 때, 변화가 있는 row 만 업데이트
+            } else if (isRenewal == SynchronizeState.UPDATE.code) {
+                for (SynonymCategory synonymCategory : synonymCategoryList) {
+                    if (changedCatNo.contains(synonymCategory.getCatNo())) {
+                        synonymCategoryRepository.save(synonymCategory);
+                    }
+                }
+            }
             responseMap.put("ok", "Synchronized");
             return new ResponseEntity<>(responseMap, HttpStatus.OK);
